--- KtistaClient/app/src/main/java/com/lkorasik/ktistaclient/ImageCaptureTypes.kt
@@ -1,6 +1,0 @@
-package com.lkorasik.ktistaclient
-
-enum class ImageCaptureTypes {
-    TAKE,
-    SELECT
-}
