+++ KtistaClient/app/src/main/java/com/lkorasik/ktistaclient/ui/helper/ImageCaptureTypes.kt
@@ -1,0 +1,6 @@
+package com.lkorasik.ktistaclient.ui.helper
+
+enum class ImageCaptureTypes {
+    CAMERA,
+    GALLERY
+}
